+++ mse-20190531/core/Models/InitializeServiceLinkRoleRequest.cs
@@ -1,0 +1,38 @@
+// This file is auto-generated, don't edit it. Thanks.
+
+using System;
+using System.Collections.Generic;
+using System.IO;
+
+using Tea;
+
+namespace AlibabaCloud.SDK.Mse20190531.Models
+{
+    public class InitializeServiceLinkRoleRequest : TeaModel {
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>zh</para>
+        /// </summary>
+        [NameInMap("AcceptLanguage")]
+        [Validation(Required=false)]
+        public string AcceptLanguage { get; set; }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>AliyunServiceRoleForMSE</para>
+        /// </summary>
+        [NameInMap("RoleName")]
+        [Validation(Required=false)]
+        public string RoleName { get; set; }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>&quot;&quot;</para>
+        /// </summary>
+        [NameInMap("Token")]
+        [Validation(Required=false)]
+        public string Token { get; set; }
+
+    }
+
+}