+++ mse-20190531/core/Models/InitializeServiceLinkRoleResponseBody.cs
@@ -1,0 +1,76 @@
+// This file is auto-generated, don't edit it. Thanks.
+
+using System;
+using System.Collections.Generic;
+using System.IO;
+
+using Tea;
+
+namespace AlibabaCloud.SDK.Mse20190531.Models
+{
+    public class InitializeServiceLinkRoleResponseBody : TeaModel {
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>200</para>
+        /// </summary>
+        [NameInMap("Code")]
+        [Validation(Required=false)]
+        public string Code { get; set; }
+
+        [NameInMap("Data")]
+        [Validation(Required=false)]
+        public InitializeServiceLinkRoleResponseBodyData Data { get; set; }
+        public class InitializeServiceLinkRoleResponseBodyData : TeaModel {
+            /// <summary>
+            /// <b>Example:</b>
+            /// <para>ram:CreateServiceLinkedRole</para>
+            /// </summary>
+            [NameInMap("RequiredPermission")]
+            [Validation(Required=false)]
+            public string RequiredPermission { get; set; }
+
+            /// <summary>
+            /// <b>Example:</b>
+            /// <para>AliyunServiceRoleForMSE</para>
+            /// </summary>
+            [NameInMap("RoleName")]
+            [Validation(Required=false)]
+            public string RoleName { get; set; }
+
+            /// <summary>
+            /// <b>Example:</b>
+            /// <para>mse.aliyuncs.com</para>
+            /// </summary>
+            [NameInMap("ServiceName")]
+            [Validation(Required=false)]
+            public string ServiceName { get; set; }
+
+        }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>OK</para>
+        /// </summary>
+        [NameInMap("Message")]
+        [Validation(Required=false)]
+        public string Message { get; set; }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>A5E7D4E3-D30C-56C1-817F-F2B8CE6BXXXX</para>
+        /// </summary>
+        [NameInMap("RequestId")]
+        [Validation(Required=false)]
+        public string RequestId { get; set; }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>true</para>
+        /// </summary>
+        [NameInMap("Success")]
+        [Validation(Required=false)]
+        public bool? Success { get; set; }
+
+    }
+
+}