+++ mse-20190531/core/Models/ListListenersByConfigRequest.cs
@@ -34,6 +34,28 @@ namespace AlibabaCloud.SDK.Mse20190531.Models
         [Validation(Required=false)]
         public string DataId { get; set; }
 
+        [NameInMap("ExtGrayRules")]
+        [Validation(Required=false)]
+        public List<ListListenersByConfigRequestExtGrayRules> ExtGrayRules { get; set; }
+        public class ListListenersByConfigRequestExtGrayRules : TeaModel {
+            [NameInMap("GrayRule")]
+            [Validation(Required=false)]
+            public string GrayRule { get; set; }
+
+            [NameInMap("GrayRuleName")]
+            [Validation(Required=false)]
+            public string GrayRuleName { get; set; }
+
+            [NameInMap("GrayRulePriority")]
+            [Validation(Required=false)]
+            public int? GrayRulePriority { get; set; }
+
+            [NameInMap("GrayRuleType")]
+            [Validation(Required=false)]
+            public string GrayRuleType { get; set; }
+
+        }
+
         /// <summary>
         /// <para>The name of the group.</para>
         /// <para>This parameter is required.</para>